HAR builder will now add ALL cookies
The current implementation of netexport did not include javascript
created session cookies in the export.  Although it breaks the current
protocol, we wanted a report that included ALL of the cookies.
Exporting data will now include a list of all of the cookies that the
browser knows about in the HAR file.